OBD2: add Mode 22 (UDS) support — 16-bit PIDs end-to-end
Mode 22 ("Read Data By Identifier") is the modern OBD-2 diagnostic
service used by Ford, GM, VW, BMW, and newer Toyotas (often supported
alongside Mode 21 on the same engine ECU). 16-bit PIDs unlock the
entire UDS namespace — typically thousands of values per vehicle vs.
the ~80 standard J1979 PIDs.

Protocol layer (obd2.c):
 - _send_pid_request now branches on service: 8-bit PID framing
   (length 2) for Mode 01/21, 16-bit PID framing (length 3, two PID
   bytes) for Mode 22.
 - _process_full_message has a dedicated Mode 22 path: extracts the
   2-byte PID echo, slices payload from msg[3], runs the existing
   packed / single-value decoder. Test intercept + freshness update
   + response-triggered dispatch all share the same code as Mode 21.
 - ISO-TP multi-frame RX already supports Mode 22 responses by
   coincidence (it's protocol-agnostic).

Type widening:
 - obd2_pid_def_t.pid: uint8_t -> uint16_t. All existing initializers
   compile unchanged (values fit).
 - obd2_poll_state_t.pid: uint8_t -> uint16_t.
 - obd2_pid_find / obd2_pid_find_svc / obd2_custom_find_svc /
   obd2_custom_add: pid params widened.
 - signal_row_t.parent_pid (obd2_picker): widened so Mode 22 rows
   carry the full PID.
 - preconfig_item_t.obd2_pid (preset_picker): widened so long-press
   picker entries can address Mode 22 PIDs.

Encoded (service, pid) storage:
 - obd2_encode_pid -> uint32_t.
 - Three encoding ranges decoded by value:
     <= 0xFF       legacy Mode 01 bare byte
     0x100..0xFFFF (svc<<8)|pid — Mode 01/21 8-bit PID
     > 0xFFFF      (svc<<16)|pid — Mode 22 16-bit PID
 - Encoder picks the smallest valid form so existing layouts
   serialise to the same shape as before (no JSON bloat for the
   common case). Readers handle all three transparently.
 - obd2_start / obd2_get_enabled / obd2_set_enabled signatures
   uint16_t -> uint32_t.
 - ecu_preset_save/read_obd2_pids: same widening + readers accept
   values up to 0xFFFFFF.
 - layout_manager._load_signals + build_json: handles uint32 form.

UI:
 - Web Custom PIDs editor: Mode 22 dropdown option no longer
   "(not wired yet)"; PID field accepts up to 4 hex digits when
   Mode 22 selected. Table cell shows 0x115C-style 4-digit PIDs
   for Mode 22 entries.
 - obd2_picker row label format: M22 entries show "(M22·0x115C)"
   with full 4-digit PID instead of truncating.
 - preset_picker preview footer: same 4-digit format for Mode 22.

Long-press picker DOES expose Mode 22 PIDs now (preconfig_item_t
holds them in uint16_t). User can pick "MY_FORD_PID" (or whatever
they named a Mode 22 custom PID) from the long-press picker just
like any other channel.

+++ main/can/obd2_pids.c
@@ -232,7 +232,7 @@ const obd2_pid_def_t OBD2_PIDS[] = {
 
 const int OBD2_PIDS_COUNT = (int)(sizeof(OBD2_PIDS) / sizeof(OBD2_PIDS[0]));
 
-const obd2_pid_def_t *obd2_pid_find_svc(uint8_t service, uint8_t pid)
+const obd2_pid_def_t *obd2_pid_find_svc(uint8_t service, uint16_t pid)
 {
     if (service == 0) service = 0x01;
     /* Built-in first. */
@@ -249,7 +249,7 @@ const obd2_pid_def_t *obd2_pid_find_svc(uint8_t service, uint8_t pid)
     return obd2_pid_find(pid);
 }
 
-const obd2_pid_def_t *obd2_pid_find(uint8_t pid)
+const obd2_pid_def_t *obd2_pid_find(uint16_t pid)
 {
     for (int i = 0; i < OBD2_PIDS_COUNT; i++) {
         if (OBD2_PIDS[i].pid == pid) return &OBD2_PIDS[i];
